requirements.txt
numpy>=1.24.0
pillow>=10.0.0
pdf2image>=1.16.0
streamlit>=1.28.0
pytesseract>=0.3.10
opencv-python>=4.8.0
watchdog>=3.0.0
imagehash>=4.3.1
pandas>=2.0.0